+++ tests/Deedle.MathNetNumerics.Tests/Stats.fs
@@ -1,13 +1,13 @@
 #if INTERACTIVE
 #I "../../bin/netstandard2.0"
 #load "Deedle.fsx"
-#load "Deedle.Math.fsx"
+#load "Deedle.MathNetNumerics.fsx"
 #r "../../packages/NUnit/lib/net45/nunit.framework.dll"
 #r "../../packages/FsCheck/lib/net452/FsCheck.dll"
 #r "../../packages/FsUnit/lib/net45/FsUnit.NUnit.dll"
 #load "../Common/FsUnit.fs"
 #else
-module Deedle.Math.Tests.Stats
+module Deedle.MathNetNumerics.Tests.Stats
 #endif
 
 open System
@@ -16,7 +16,7 @@ open NUnit.Framework
 open FsCheck
 open Deedle
 open Deedle.Internal
-open Deedle.Math
+open Deedle.MathNetNumerics
 open MathNet.Numerics.Statistics
 
 let stockPrices = Frame.ReadCsv(__SOURCE_DIRECTORY__ + "/data/stocks_weekly.csv") |> Frame.indexRowsDate "Dates"